Add skeletons for remaning read builders. Move EnumType to models WIP.

+++ server/prisma-rs/query-engine/connectors/sql-connector/src/query_builder/read/many_related_records/base_query.rs
@@ -13,7 +13,7 @@ pub struct ManyRelatedRecordsBaseQuery<'a> {
     pub is_reverse_order: bool,
     pub condition: ConditionTree<'static>,
     pub cursor: ConditionTree<'static>,
-    pub window_limits: (u32, u32),
+    pub window_limits: (i64, i64),
     pub skip_and_limit: SkipAndLimit,
 }
